npu/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
transformers==4.46.2
accelerate==1.1.1
optimum-intel[openvino]==1.20.0
openvino==2024.4.0
openvino-genai==2024.4.0
python-dotenv==1.0.1
